Added fusion function in 2d landoltc

+++ pandora_vision_landoltc/include/pandora_vision_landoltc/landoltc_parameters.h
@@ -45,8 +45,8 @@ namespace pandora_vision
   struct LandoltcParameters
   {
     //!< Threshold parameters
-    static int gradientThreshold;
-    static int centerThreshold;
+    static double gradientThreshold;
+    static double centerThreshold;
     static double huMomentsPrec;
     static int adaptiveThresholdSubtractSize;
